# Conflicts:
#	src/main/java/tyut/selab/utils/Result.java

+++ src/main/java/tyut/selab/userservice/service/UserService.java
@@ -26,7 +26,7 @@ public interface UserService {
      * @param userId
      * @return
      */
-    public UserVo selectByUserId(Integer userId);
+    public UserVo selectByUserId(Long userId);
 
     /**
      *  通过id删除用户
